Update and rename testing_with_pytest.py to test_with_pytest.py
Every file and function that has the keyword assert in the file needs to start with the word "test" in order for pytest to detect the file and collect any items in the run, so the file is renamed to "test_with_pytest".py. If the file does not start with the word "test", running this file in the Terminal will display a message that does not show an error or a success. 

To run a file in the Terminal, type the word "pytest". A few runs will show an error if the two random numbers a and b that are being multiplied does equal or not equal integer c.

Also, a few edits have been made to the script. pytest needs to be installed with "pip install pytest" if it is not on the machine and it does not to be imported from the top of the script. The declared integers are in the function now and no parameters are in the script now. The function has been renamed to "test_check_multiplication()" for pytest to detect the function.