Update Dockerfile
Fixed syntax issues. -y means yes and should come after installing the name of the package to ask, whether it should be downloaded. Most packages require a prompt yes or no to download the package, and including -y automatically downloads the packages. This feature is useful if a script is being automated or a Dockerfile is being ran.